move custom logic to client
