+++ matter_server/common/helpers/util.py
@@ -1,6 +1,7 @@
 """Utils for Matter server (and client)."""
 from __future__ import annotations
 
+import base64
 from base64 import b64decode
 import binascii
 from dataclasses import MISSING, asdict, fields, is_dataclass
@@ -10,6 +11,7 @@ from functools import cache
 from importlib.metadata import PackageNotFoundError, version as pkg_version
 import logging
 import platform
+import socket
 from types import NoneType, UnionType
 from typing import (
     TYPE_CHECKING,
@@ -219,8 +221,7 @@ def parse_value(
     ):
         return uint(value)
     if value_type is float32 and (
-        isinstance(value, float)
-        or isinstance(value, int)
+        isinstance(value, (float, int))
         or (isinstance(value, str) and value.isnumeric())
     ):
         return float32(value)
@@ -291,3 +292,33 @@ def chip_core_version() -> str:
         # TODO: Fix this once we can install our own wheels on macos.
         return chip_clusters_version()
     return package_version(CHIP_CORE_PKG_NAME)
+
+
+def convert_hex_string(hex_str: str | bytes) -> str:
+    """Convert hex string received from the sdk to a regular (unicode) string."""
+    if isinstance(hex_str, str):
+        # note that the bytes string can be optionally base64 encoded
+        # when we send it back and forth over our api
+        hex_str = base64.b64decode(hex_str)
+
+    return "".join(f"{byte:02x}" for byte in hex_str)
+
+
+def convert_mac_address(hex_mac: str | bytes) -> str:
+    """Convert hexadecimal MAC received from the sdk to a regular mac-address string."""
+    if isinstance(hex_mac, str):
+        # note that the bytes string can be optionally base64 encoded
+        hex_mac = base64.b64decode(hex_mac)
+
+    return ":".join("{:02x}".format(byte) for byte in hex_mac)  # pylint: disable=C0209
+
+
+def convert_ip_address(hex_ip: str | bytes, ipv6: bool = False) -> str:
+    """Convert hexadecimal IP received from the sdk to a regular IP string."""
+    if isinstance(hex_ip, str):
+        # note that the bytes string can be optionally base64 encoded
+        hex_ip = base64.b64decode(hex_ip)
+    if ipv6:
+        hex_str = "".join(f"{byte:02x}" for byte in hex_ip)
+        return ":".join(hex_str[i : i + 4] for i in range(0, len(hex_str), 4))
+    return socket.inet_ntoa(hex_ip)
